apps/pt_losses/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pt-losses"
version = "0.1.0"
description = "Herramienta modular para calculo de perdidas de postensado orientada a RFEM 6."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
rfem = [
    "dlubal.api",
]

[project.scripts]
pt-losses = "pt_losses.cli.main:run"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"pt_losses.gui" = ["assets/*"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
